requirements.txt
tabula-py
openpyxl
pandas
PySimpleGUI
PyMuPDF
Pillow